+++ libraries/74HC590/CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2026-01-21
+- add DEV_54HC590 derived class.
+- add DEV_54HC590 example
+- update GitHub actions
+- minor edits
+
 ## [0.1.0] - 2024-04-30
 - initial version
-
-